feat: 라우팅 방식을 수정했다. SearchAllpage에서 사용하는 input과 검색어 처리 로직들이 이후 SearchUserPage, SearchPostPage와 중복되기 때문에 nested route를 사용해 중복되는 부분을 SearchPage로 분리했다.
