+++ src/App.js
@@ -12,9 +12,10 @@ import SearchPage from '@pages/SearchPage';
 import SearchPostPage from '@pages/SearchPostPage';
 import SearchUserPage from '@pages/SearchUserPage';
 import SignupPage from '@pages/SignupPage';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import ContextProvider from '@contexts/ContextProvider';
 import InterestedChannelModal from '@components/InterestedChannelModal';
+import SearchAllPage from '@pages/SearchAllPage';
 
 function App() {
   return (
@@ -25,9 +26,12 @@ function App() {
           <Route path="/" element={<Navbar />}>
             <Route index element={<HomePage />} />
             <Route path="alram/:user" element={<AlramPage />} />
-            <Route path="search/all" element={<SearchPage />} />
-            <Route path="search/user" element={<SearchUserPage />} />
-            <Route path="search/post" element={<SearchPostPage />} />
+            <Route path="search" element={<SearchPage />}>
+              <Route path="" element={<Navigate to="all" replace />} />
+              <Route path="all" element={<SearchAllPage />} />
+              <Route path="user" element={<SearchUserPage />} />
+              <Route path="post" element={<SearchPostPage />} />
+            </Route>
             <Route path="channel/:channelId" element={<ChannelPage />} />
             <Route path="channel/categories" element={<CategoriesPage />} />
             <Route path="posts/write" element={<PostWritePage />} />
