+++ src/pages/SearchPage.js
@@ -1,39 +1,19 @@
-import Card from '@components/Card';
-import Header from '@components/Header';
 import styled from '@emotion/styled';
 import useAsyncFn from '@hooks/useAsyncFn';
-import { Button, IconButton, TextField } from '@mui/material';
+import { IconButton, TextField } from '@mui/material';
 import SearchRoundedIcon from '@mui/icons-material/SearchRounded';
 import { searchAll } from '@utils/search';
 import React, { useEffect, useRef, useState } from 'react';
-import { Link, useSearchParams } from 'react-router-dom';
-import { COLOR_SIGNATURE } from '@utils/color';
-import List from '@components/List';
+import { Outlet, useSearchParams } from 'react-router-dom';
 
-const HeaderBox = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-`;
-const ListWrapper = styled.div`
-  flex-grow: 1;
-`;
-const HeaderWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
 const Container = styled.div`
   height: 100%;
   display: grid;
-  grid-template-rows: 5rem 1fr 1fr;
+  grid-template-rows: 5rem 1fr;
   grid-template-columns: 100%;
   justify-items: center;
 `;
-const RouteLink = styled(Link)`
-  text-decoration: none;
-  color: ${COLOR_SIGNATURE};
-`;
+
 function SearchPage() {
   const inputRef = useRef(null);
   const [searchParam, setSearchParam] = useSearchParams();
@@ -61,8 +41,8 @@ function SearchPage() {
 
   useEffect(() => {
     const keyword = searchParam.get('k');
-    if (keyword === null) return;
     inputRef.current.value = keyword;
+    if (keyword === null) return;
     isValidKeyword(keyword) && searchWithKeword(keyword);
   }, [searchParam, searchWithKeword]);
 
@@ -78,16 +58,6 @@ function SearchPage() {
     isValidKeyword(inputRef.current.value) &&
       setSearchParam({ k: inputRef.current.value });
   };
-  const filterUser = () => {
-    if (!Array.isArray(searchResult.value)) return [];
-    return searchResult.value.filter((data) => data.fullName);
-  };
-  const filterPost = () => {
-    if (!Array.isArray(searchResult.value)) return [];
-    return searchResult.value
-      .filter((data) => data.title)
-      .sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt));
-  };
   const InputProps = {
     endAdornment: (
       <IconButton onClick={handleSearchClick}>
@@ -96,7 +66,6 @@ function SearchPage() {
     ),
     style: { fontSize: 20 },
   };
-
   return (
     <Container>
       <TextField
@@ -106,58 +75,7 @@ function SearchPage() {
         onKeyDown={handleKeyDown}
         {...inputError}
       />
-      {!searchResult.isLoading && searchParam.get('k') && (
-        <>
-          <HeaderBox>
-            <HeaderWrapper>
-              <Header level={1} strong>
-                사용자
-              </Header>
-              <Button variant="text" color="primary">
-                <RouteLink to={`/search/user?k=${searchParam.get('k')}`}>
-                  모두 보기
-                </RouteLink>
-              </Button>
-            </HeaderWrapper>
-            <ListWrapper>
-              <List
-                items={filterUser()}
-                keyName="_id"
-                alt="키워드에 해당하는 결과가 없습니다"
-                limit={3}
-                gap="8px"
-                frame={<Card.User />}
-              />
-            </ListWrapper>
-          </HeaderBox>
-          <HeaderBox>
-            <HeaderWrapper>
-              <Header level={1} strong>
-                포스트
-              </Header>
-              <Button variant="text" color="primary">
-                <RouteLink to={`/search/post?k=${searchParam.get('k')}`}>
-                  모두 보기
-                </RouteLink>
-              </Button>
-            </HeaderWrapper>
-            <ListWrapper>
-              <List
-                items={filterPost()}
-                keyName="_id"
-                alt="키워드에 해당하는 결과가 없습니다"
-                limit={3}
-                gap="8px"
-                frame={<Card.Post />}
-              />
-            </ListWrapper>
-          </HeaderBox>
-        </>
-      )}
-      {!searchParam.get('k') && (
-        <Header margin="10% 0">검색어를 입력해 파티를 찾아보세요!</Header>
-      )}
-      {searchResult.isLoading && <Header margin="10% 0">로딩중!</Header>}
+      <Outlet context={{ ...searchResult }} />
     </Container>
   );
 }
